+++ typedoc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  out: "public/docs",
+  exclude: ["**/*+(index).ts", "**/__tests__/**/*"],
+};