+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.0
+- Added `parseOPM` [PR#2](https://github.com/digital-sound-antiques/ym-voice/pull/2). Thanks to @dquenne
+- Fixed encode error in `ws` of OPLL.
+- Fixed encode error in `dt2` of OPM.
+- Upgraded NPM dependencies.
+- Changed minimum Node version to 18.
+
 ## 0.2.0
 - Added `toFile` method. Thanks to @dquenne